Move evaluate_commands_map out of evaluation.py
