+++ src/statue/cli/run.py
@@ -18,9 +18,10 @@ from statue.cli.util import (
     verbosity_option,
 )
 from statue.commands_map import read_commands_map
-from statue.evaluation import Evaluation, evaluate_commands_map
+from statue.evaluation import Evaluation
 from statue.exceptions import MissingConfiguration, UnknownContext
 from statue.print_util import print_boxed
+from statue.runner import evaluate_commands_map
 from statue.verbosity import is_silent
